[level 2] Title: 피보나치 수, Time: 0.28 ms, Memory: 3.74 MB -BaekjoonHub

+++ 프로그래머스/2/12945. 피보나치 수/피보나치 수.cpp
@@ -1,13 +1,10 @@
-#include <string>
-#include <iostream>
-#include <vector>
 
 using namespace std;
 
 int solution(int n) {
-    unsigned long long fNminus2 = 0;
-    unsigned long long fNminus1 = 1;
-    unsigned long long fN = 1;
+    unsigned int fNminus2 = 0;
+    unsigned int fNminus1 = 1;
+    unsigned int fN = 1;
     
     for (int i = 2; i <= n; i++)
     {
